@@ -52,8 +52,8 @@
     "nwb": "0.21.x",
     "prettier": "^1.10.2",
     "radium": "^0.21.2",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
+    "react": "^16.3.0",
+    "react-dom": "^16.3.0",
     "recompose": "^0.26.0",
     "semantic-release": "^12.4.1"
   },
