@@ -1,27 +1,27 @@
 import expect, { createSpy } from 'expect'
 import * as utils from '../src/utils'
 
-describe('utils', function() {
-  var defaultMarkup = '@[__display__](__type__:__id__)'
-  var value =
+describe('utils', () => {
+  const defaultMarkup = '@[__display__](__type__:__id__)'
+  const value =
     "Hi @[John Doe](user:johndoe), \n\nlet's add @[joe@smoe.com](email:joe@smoe.com) to this conversation..."
-  var plainText =
+  const plainText =
     "Hi John Doe, \n\nlet's add joe@smoe.com to this conversation..."
 
-  var displayTransform = function(id) {
+  const displayTransform = function(id) {
     return '<--' + id + '-->'
   }
-  var plainTextDisplayTransform =
+  const plainTextDisplayTransform =
     "Hi <--johndoe-->, \n\nlet's add <--joe@smoe.com--> to this conversation..."
 
-  describe('#spliceString', function() {
-    it('should replace the substring between start and end with the provided insertion', function() {
+  describe('#spliceString', () => {
+    it('should replace the substring between start and end with the provided insertion', () => {
       expect(utils.spliceString('012345678', 1, 4, 'xx')).toEqual('0xx45678')
     })
   })
 
-  describe('#getPositionOfCapturingGroup', function() {
-    var testData = {
+  describe('#getPositionOfCapturingGroup', () => {
+    const testData = {
       '@[__display__](__id__)': { display: 0, id: 1, type: null },
       '@[__display__](__type__:__id__)': { display: 0, id: 2, type: 1 },
       '@(__type__:__id__)': { display: 1, id: 1, type: 0 },
@@ -31,8 +31,8 @@ describe('utils', function() {
     }
 
     Object.keys(testData).forEach(key => {
-      var markup = key
-      var positions = testData[key]
+      const markup = key
+      const positions = testData[key]
 
       it(
         'should return ' +
@@ -40,7 +40,7 @@ describe('utils', function() {
           ' for the `display` position in markup `' +
           markup +
           '`',
-        function() {
+        () => {
           expect(utils.getPositionOfCapturingGroup(markup, 'display')).toEqual(
             positions.display
           )
@@ -53,7 +53,7 @@ describe('utils', function() {
           ' for the `id` position in markup `' +
           markup +
           '`',
-        function() {
+        () => {
           expect(utils.getPositionOfCapturingGroup(markup, 'id')).toEqual(
             positions.id
           )
@@ -66,7 +66,7 @@ describe('utils', function() {
           ' for the `type` position in markup `' +
           markup +
           '`',
-        function() {
+        () => {
           expect(utils.getPositionOfCapturingGroup(markup, 'type')).toEqual(
             positions.type
           )
@@ -75,13 +75,13 @@ describe('utils', function() {
     })
   })
 
-  describe('#iterateMentionsMarkup', function() {
-    it('should call the `markupIteratee` for every markup occurrence', function() {
-      var markupIteratee = createSpy()
+  describe('#iterateMentionsMarkup', () => {
+    it('should call the `markupIteratee` for every markup occurrence', () => {
+      const markupIteratee = createSpy()
       utils.iterateMentionsMarkup(
         value,
         defaultMarkup,
-        function() {},
+        () => {},
         markupIteratee
       )
 
@@ -107,14 +107,43 @@ describe('utils', function() {
       )
     })
 
-    it('should call the `textIteratee` for all plain text sub string between markups', function() {
-      var textIteratee = createSpy()
+    it('should call the `markupIteratee` with the correct plain text indices when a display transform is used', () => {
+      const markupIteratee = createSpy()
       utils.iterateMentionsMarkup(
         value,
         defaultMarkup,
-        textIteratee,
-        function() {}
+        () => {},
+        markupIteratee,
+        (id, display, type) => `[${display}]`
       )
+      const plainTextWithDisplayTransform =
+        "Hi [John Doe], \n\nlet's add [joe@smoe.com] to this conversation..."
+
+      expect(markupIteratee.calls.length).toEqual(2)
+      expect(markupIteratee).toHaveBeenCalledWith(
+        '@[John Doe](user:johndoe)',
+        value.indexOf('@[John Doe](user:johndoe)'),
+        plainTextWithDisplayTransform.indexOf('[John Doe]'),
+        'johndoe',
+        '[John Doe]',
+        'user',
+        0
+      )
+      expect(markupIteratee).toHaveBeenCalledWith(
+        '@[joe@smoe.com](email:joe@smoe.com)',
+        value.indexOf('@[joe@smoe.com](email:joe@smoe.com)'),
+        plainTextWithDisplayTransform.indexOf('[joe@smoe.com]'),
+        'joe@smoe.com',
+        '[joe@smoe.com]',
+        'email',
+        value.indexOf('@[John Doe](user:johndoe)') +
+          '@[John Doe](user:johndoe)'.length
+      )
+    })
+
+    it('should call the `textIteratee` for all plain text sub string between markups', () => {
+      const textIteratee = createSpy()
+      utils.iterateMentionsMarkup(value, defaultMarkup, textIteratee, () => {})
 
       expect(textIteratee.calls.length).toEqual(3)
       expect(textIteratee).toHaveBeenCalledWith('Hi ', 0, 0)
@@ -130,12 +159,12 @@ describe('utils', function() {
       )
     })
 
-    it('should call the `markupIteratee` for every markup occurrence with display transform', function() {
-      var markupIteratee = createSpy()
+    it('should call the `markupIteratee` for every markup occurrence with display transform', () => {
+      const markupIteratee = createSpy()
       utils.iterateMentionsMarkup(
         value,
         defaultMarkup,
-        function() {},
+        () => {},
         markupIteratee,
         displayTransform
       )
@@ -162,13 +191,13 @@ describe('utils', function() {
       )
     })
 
-    it('should call the `textIteratee` for all plain text sub string between markups with display transform', function() {
-      var textIteratee = createSpy()
+    it('should call the `textIteratee` for all plain text sub string between markups with display transform', () => {
+      const textIteratee = createSpy()
       utils.iterateMentionsMarkup(
         value,
         defaultMarkup,
         textIteratee,
-        function() {},
+        () => {},
         displayTransform
       )
 
@@ -187,16 +216,20 @@ describe('utils', function() {
     })
   })
 
-  describe('#mapPlainTextIndex', function() {
-    it('should correctly calculate the index of a character in the plain text between mentions', function() {
-      var plainTextIndex = plainText.indexOf("let's add")
-      var result = utils.mapPlainTextIndex(value, defaultMarkup, plainTextIndex)
+  describe('#mapPlainTextIndex', () => {
+    it('should correctly calculate the index of a character in the plain text between mentions', () => {
+      const plainTextIndex = plainText.indexOf("let's add")
+      const result = utils.mapPlainTextIndex(
+        value,
+        defaultMarkup,
+        plainTextIndex
+      )
       expect(result).toEqual(value.indexOf("let's add"))
     })
 
-    it('should correctly calculate the index of a character in the plain text between mentions with display tranform', function() {
-      var plainTextIndex = plainTextDisplayTransform.indexOf("let's add")
-      var result = utils.mapPlainTextIndex(
+    it('should correctly calculate the index of a character in the plain text between mentions with display tranform', () => {
+      const plainTextIndex = plainTextDisplayTransform.indexOf("let's add")
+      const result = utils.mapPlainTextIndex(
         value,
         defaultMarkup,
         plainTextIndex,
@@ -206,36 +239,44 @@ describe('utils', function() {
       expect(result).toEqual(value.indexOf("let's add"))
     })
 
-    it('should correctly calculate the indices of the character in the plain text before the first mention', function() {
-      var result = utils.mapPlainTextIndex(value, defaultMarkup, 2)
+    it('should correctly calculate the indices of the character in the plain text before the first mention', () => {
+      const result = utils.mapPlainTextIndex(value, defaultMarkup, 2)
       expect(result).toEqual(2)
     })
 
-    it('should correctly calculate the index of a character in the plain text after the last mention', function() {
-      var plainTextIndex = plainText.indexOf('...')
-      var result = utils.mapPlainTextIndex(value, defaultMarkup, plainTextIndex)
+    it('should correctly calculate the index of a character in the plain text after the last mention', () => {
+      const plainTextIndex = plainText.indexOf('...')
+      const result = utils.mapPlainTextIndex(
+        value,
+        defaultMarkup,
+        plainTextIndex
+      )
       expect(result).toEqual(value.indexOf('...'))
     })
 
-    it('should correctly calculate the index of the first plain text character after a mention', function() {
-      var plainTextIndex = plainText.indexOf(',') // first char after John Doe mention
-      var result = utils.mapPlainTextIndex(value, defaultMarkup, plainTextIndex)
+    it('should correctly calculate the index of the first plain text character after a mention', () => {
+      const plainTextIndex = plainText.indexOf(',') // first char after John Doe mention
+      const result = utils.mapPlainTextIndex(
+        value,
+        defaultMarkup,
+        plainTextIndex
+      )
       expect(result).toEqual(value.indexOf(','))
     })
 
-    it('should return the input index if there are no mentions', function() {
-      var result = utils.mapPlainTextIndex(plainText, defaultMarkup, 10)
+    it('should return the input index if there are no mentions', () => {
+      const result = utils.mapPlainTextIndex(plainText, defaultMarkup, 10)
       expect(result).toEqual(10)
     })
 
-    it("should return the index of the corresponding markup's first character if the plain text index lies inside a mention", function() {
+    it("should return the index of the corresponding markup's first character if the plain text index lies inside a mention", () => {
       // index for first char of markup
-      var plainTextIndex = plainText.indexOf('John Doe')
-      var result = utils.mapPlainTextIndex(value, defaultMarkup, plainTextIndex)
+      let plainTextIndex = plainText.indexOf('John Doe')
+      let result = utils.mapPlainTextIndex(value, defaultMarkup, plainTextIndex)
       expect(result).toEqual(value.indexOf('@[John Doe](user:johndoe)'))
 
       // index of char inside the markup
-      var joeMarkup = '@[joe@smoe.com](email:joe@smoe.com)'
+      const joeMarkup = '@[joe@smoe.com](email:joe@smoe.com)'
       plainTextIndex = plainText.indexOf('joe@smoe.com') + 3
       result = utils.mapPlainTextIndex(value, defaultMarkup, plainTextIndex)
       expect(result).toEqual(value.indexOf(joeMarkup))
@@ -247,10 +288,10 @@ describe('utils', function() {
       expect(result).toEqual(value.indexOf(joeMarkup))
     })
 
-    it("should return the index of the corresponding markup's last character if the plain text index lies inside a mention and the `inMarkupCorrection` is set to 'END'", function() {
+    it("should return the index of the corresponding markup's last character if the plain text index lies inside a mention and the `inMarkupCorrection` is set to 'END'", () => {
       // index for first char of markup
-      var plainTextIndex = plainText.indexOf('John Doe')
-      var result = utils.mapPlainTextIndex(
+      let plainTextIndex = plainText.indexOf('John Doe')
+      let result = utils.mapPlainTextIndex(
         value,
         defaultMarkup,
         plainTextIndex,
@@ -259,7 +300,7 @@ describe('utils', function() {
       expect(result).toEqual(value.indexOf('@[John Doe](user:johndoe)'))
 
       // index of char inside the markup
-      var joeMarkup = '@[joe@smoe.com](email:joe@smoe.com)'
+      const joeMarkup = '@[joe@smoe.com](email:joe@smoe.com)'
       plainTextIndex = plainText.indexOf('joe@smoe.com') + 3
       result = utils.mapPlainTextIndex(
         value,
@@ -281,10 +322,10 @@ describe('utils', function() {
       expect(result).toEqual(value.indexOf(joeMarkup) + joeMarkup.length)
     })
 
-    it("should return `null` if `inMarkupCorrection` is set to 'NULL'", function() {
+    it("should return `null` if `inMarkupCorrection` is set to 'NULL'", () => {
       // index of char inside the markup
-      var plainTextIndex = plainText.indexOf('joe@smoe.com') + 3
-      var result = utils.mapPlainTextIndex(
+      const plainTextIndex = plainText.indexOf('joe@smoe.com') + 3
+      const result = utils.mapPlainTextIndex(
         value,
         defaultMarkup,
         plainTextIndex,
@@ -293,18 +334,23 @@ describe('utils', function() {
       expect(result).toEqual(null)
     })
 
-    it("should return the index of the corresponding markup's first character if the plain text index lies inside a mention with display transform", function() {
+    it("should return the index of the corresponding markup's first character if the plain text index lies inside a mention with display transform", () => {
       // index of char inside the markup
-      var joeMarkup = '@[joe@smoe.com](email:joe@smoe.com)'
-      var plainTextIndex = plainTextDisplayTransform.indexOf('joe@smoe.com') + 3
-      var result = utils.mapPlainTextIndex(value, defaultMarkup, plainTextIndex)
+      const joeMarkup = '@[joe@smoe.com](email:joe@smoe.com)'
+      const plainTextIndex =
+        plainTextDisplayTransform.indexOf('joe@smoe.com') + 3
+      const result = utils.mapPlainTextIndex(
+        value,
+        defaultMarkup,
+        plainTextIndex
+      )
       expect(result).toEqual(value.indexOf(joeMarkup))
     })
 
-    it('should return the correctly mapped caret position at the end of the string after a mention', function() {
-      var value = 'Hi @[John Doe](user:johndoe)'
-      var plainText = 'Hi John Doe'
-      var result = utils.mapPlainTextIndex(
+    it('should return the correctly mapped caret position at the end of the string after a mention', () => {
+      const value = 'Hi @[John Doe](user:johndoe)'
+      const plainText = 'Hi John Doe'
+      const result = utils.mapPlainTextIndex(
         value,
         defaultMarkup,
         plainText.length,
@@ -314,9 +360,9 @@ describe('utils', function() {
     })
   })
 
-  describe('#findStartOfMentionInPlainText', function() {
-    it("should return the index of the mention's first char in the plain text if the passed index lies inside a mention", function() {
-      var result = utils.findStartOfMentionInPlainText(
+  describe('#findStartOfMentionInPlainText', () => {
+    it("should return the index of the mention's first char in the plain text if the passed index lies inside a mention", () => {
+      const result = utils.findStartOfMentionInPlainText(
         value,
         defaultMarkup,
         plainText.indexOf('Doe')
@@ -324,8 +370,8 @@ describe('utils', function() {
       expect(result).toEqual(plainText.indexOf('John Doe'))
     })
 
-    it('should return `undefined`, if it does not lie inside a mention', function() {
-      var result = utils.findStartOfMentionInPlainText(
+    it('should return `undefined`, if it does not lie inside a mention', () => {
+      const result = utils.findStartOfMentionInPlainText(
         value,
         defaultMarkup,
         plainText.indexOf('add')
@@ -333,8 +379,8 @@ describe('utils', function() {
       expect(result).toEqual(undefined)
     })
 
-    it("should return the index of the mention's first char if that one is the probe value", function() {
-      var result = utils.findStartOfMentionInPlainText(
+    it("should return the index of the mention's first char if that one is the probe value", () => {
+      const result = utils.findStartOfMentionInPlainText(
         value,
         defaultMarkup,
         plainText.indexOf('John')
@@ -343,10 +389,10 @@ describe('utils', function() {
     })
   })
 
-  describe('#applyChangeToValue', function() {
-    it('should correctly add a character at the end, beginning, and in the middle of text', function() {
-      var changed = 'S' + plainText
-      var result = utils.applyChangeToValue(
+  describe('#applyChangeToValue', () => {
+    it('should correctly add a character at the end, beginning, and in the middle of text', () => {
+      let changed = 'S' + plainText
+      let result = utils.applyChangeToValue(
         value,
         defaultMarkup,
         changed,
@@ -382,11 +428,11 @@ describe('utils', function() {
       )
     })
 
-    it('should correctly delete single characters and ranges of selected text', function() {
+    it('should correctly delete single characters and ranges of selected text', () => {
       // delete "i"
-      var changed =
+      let changed =
         "H John Doe, \n\nlet's add joe@smoe.com to this conversation..."
-      var result = utils.applyChangeToValue(
+      let result = utils.applyChangeToValue(
         value,
         defaultMarkup,
         changed,
@@ -413,10 +459,10 @@ describe('utils', function() {
       )
     })
 
-    it('should correctly add ranges of pasted text and replace the selected range with the new range', function() {
+    it('should correctly add ranges of pasted text and replace the selected range with the new range', () => {
       // add range
-      var changed = plainText.replace('add', 'add add')
-      var result = utils.applyChangeToValue(
+      let changed = plainText.replace('add', 'add add')
+      let result = utils.applyChangeToValue(
         value,
         defaultMarkup,
         changed,
@@ -439,11 +485,11 @@ describe('utils', function() {
       expect(result).toEqual(value.replace('add', 'remove'))
     })
 
-    it('should remove mentions markup contained in deleted text ranges', function() {
+    it('should remove mentions markup contained in deleted text ranges', () => {
       // delete without a range selection
-      var changed =
+      let changed =
         "Hi John Do, \n\nlet's add joe@smoe.com to this conversation..."
-      var result = utils.applyChangeToValue(
+      let result = utils.applyChangeToValue(
         value,
         defaultMarkup,
         changed,
@@ -478,11 +524,11 @@ describe('utils', function() {
       )
     })
 
-    it('should correctly add a new character after a mention at the end of the string', function() {
-      var value = 'Hi @[John Doe](user:johndoe)'
-      var changed = 'Hi John Doe,'
+    it('should correctly add a new character after a mention at the end of the string', () => {
+      const value = 'Hi @[John Doe](user:johndoe)'
+      const changed = 'Hi John Doe,'
 
-      var result = utils.applyChangeToValue(
+      const result = utils.applyChangeToValue(
         value,
         defaultMarkup,
         changed,
@@ -493,10 +539,10 @@ describe('utils', function() {
       expect(result).toEqual('Hi @[John Doe](user:johndoe),')
     })
 
-    it('should support deletion of whole words (Alt + Backspace) and whole lines (Cmd + Backspace)', function() {
-      var changed = plainText.replace('add', '')
+    it('should support deletion of whole words (Alt + Backspace) and whole lines (Cmd + Backspace)', () => {
+      const changed = plainText.replace('add', '')
 
-      var result = utils.applyChangeToValue(
+      const result = utils.applyChangeToValue(
         value,
         defaultMarkup,
         changed,
@@ -509,10 +555,10 @@ describe('utils', function() {
       )
     })
 
-    it('should support deletion to the right using Del key', function() {
-      var changed = plainText.replace('add', 'dd')
+    it('should support deletion to the right using Del key', () => {
+      const changed = plainText.replace('add', 'dd')
 
-      var result = utils.applyChangeToValue(
+      const result = utils.applyChangeToValue(
         value,
         defaultMarkup,
         changed,
@@ -525,9 +571,9 @@ describe('utils', function() {
       )
     })
 
-    it('should support deletion to the right using Del key when using the displayTransform option', function() {
-      var changed = plainTextDisplayTransform.replace('add', 'dd')
-      var result = utils.applyChangeToValue(
+    it('should support deletion to the right using Del key when using the displayTransform option', () => {
+      const changed = plainTextDisplayTransform.replace('add', 'dd')
+      const result = utils.applyChangeToValue(
         value,
         defaultMarkup,
         changed,
@@ -541,8 +587,8 @@ describe('utils', function() {
       )
     })
 
-    it('should correctly handle text auto-correction', function() {
-      var result = utils.applyChangeToValue(
+    it('should correctly handle text auto-correction', () => {
+      const result = utils.applyChangeToValue(
         'ill',
         defaultMarkup,
         "I'll",
@@ -554,9 +600,9 @@ describe('utils', function() {
     })
   })
 
-  describe('#getMentions', function() {
-    it('should return an array of all mentions in the provided value', function() {
-      var mentions = utils.getMentions(value, defaultMarkup)
+  describe('#getMentions', () => {
+    it('should return an array of all mentions in the provided value', () => {
+      const mentions = utils.getMentions(value, defaultMarkup)
       expect(mentions).toEqual([
         {
           id: 'johndoe',
@@ -575,8 +621,8 @@ describe('utils', function() {
       ])
     })
 
-    it('should take into account the displayTransform if passed', function() {
-      var mentions = utils.getMentions(value, defaultMarkup, displayTransform)
+    it('should take into account the displayTransform if passed', () => {
+      const mentions = utils.getMentions(value, defaultMarkup, displayTransform)
       expect(mentions).toEqual([
         {
           id: 'johndoe',
@@ -598,12 +644,12 @@ describe('utils', function() {
 
   describe('#getEndOfLastMention', () => {
     it('should return the end index of the last mention in the plain text', () => {
-      var index = utils.getEndOfLastMention(value, defaultMarkup)
+      const index = utils.getEndOfLastMention(value, defaultMarkup)
       expect(index).toEqual(37)
     })
 
     it('should take into account the displayTransform', () => {
-      var index = utils.getEndOfLastMention(
+      const index = utils.getEndOfLastMention(
         value,
         defaultMarkup,
         displayTransform
@@ -612,7 +658,7 @@ describe('utils', function() {
     })
 
     it('should return 0 if there is no mention', () => {
-      var index = utils.getEndOfLastMention(
+      const index = utils.getEndOfLastMention(
         'No mentions to be found here',
         defaultMarkup
       )
